Migrate the test mod to mojang mappings
A lot of this had to be edited, as Loom's migrateMappings command still produces weirdness.

RichTranslations is broken for some reason. Text seems to be serialized correctly, but the color is missing.

Biggest problem with this the migration is the fact that owo-lib uses the name `Component` for, well, its UI-components. This clashes directly with what was formerly known as `Text`, or `Components`.